Update README.md
This project is about printing the “CALENDAR” for a particular year for all the twelve months in that year. Here in this program a particular year is given as the input. Then it is checked whether the given year is leap year or not. If the given year is a leap year then the number of days in the February month becomes. Or else the number of days in the February month becomes 28. Then using a formula the first day of the first month is found out by using the “value”. This formula includes three different steps. Then using the value  obtained and the number of days in the first  month  a function is called. This function is named as “month”. In this function the days in the month are printed accordingly. Then the “value” is modified again according to the number of days in the next month to be printed. Similarly this function “month” is called twelve times to print the twelve months in the given year. Thus the calendar is being printed successfully.